requirements.txt
numpy>=1.24.0
plyfile>=1.0.0
pygltflib>=1.16.0
click>=8.0.0
Pillow>=10.0.0
meshoptimizer>=0.19.0

